More exceptions added and table shown in the entry page while the column renderers show the free text view
